requirements.txt
llama-cpp-python
psutil
flask
